Extract 2025 match preview component
